+++ AgroVA.Domain/Entities/Load.cs
@@ -6,11 +6,11 @@ namespace AgroVA.Domain.Entities
     public sealed class Load : EntityBase
     {
         public DateTime Timestamp { get; private set; }
-        public float? GreenWeight { get; private set; }
-        public float? DryWeight { get; private set; }
+        public decimal? GreenWeight { get; private set; }
+        public decimal? DryWeight { get; private set; }
         public int Register { get; private set; }
-        public float? Price { get; private set; }
-        public float? WholePercent { get; private set; }
+        public decimal? Price { get; private set; }
+        public decimal? WholePercent { get; private set; }
         public long Invoice { get; private set; }
 
         public int FarmerId { get; set; }
@@ -19,26 +19,26 @@ namespace AgroVA.Domain.Entities
         public int HarvestId { get; set; }
         public Harvest Harvest { get; set; }
 
-        public Load(int id, DateTime timestamp, float? greenWeight, float? dryWeight, int register, float? price, float? wholePercent, long invoice)
+        public Load(int id, DateTime timestamp, decimal? greenWeight, decimal? dryWeight, int register, decimal? price, decimal? wholePercent, long invoice)
         {
             DomainExceptionValidation.When(id < 1, LoadMessage.InvalidId);
             Id = id;
             ValidateDomain(timestamp, greenWeight, dryWeight, register, price, wholePercent, invoice);
         }
 
-        public Load(DateTime timestamp, float? greenWeight, float? dryWeight, int register, float? price, float? wholePercent, long invoice)
+        public Load(DateTime timestamp, decimal? greenWeight, decimal? dryWeight, int register, decimal? price, decimal? wholePercent, long invoice)
         {
             ValidateDomain(timestamp, greenWeight, dryWeight, register, price, wholePercent, invoice);
         }
 
-        public void Update(DateTime timestamp, float? greenWeight, float? dryWeight, int register, float? price, float wholePercent, long invoice, int farmerId, int harvestId)
+        public void Update(DateTime timestamp, decimal? greenWeight, decimal? dryWeight, int register, decimal? price, decimal wholePercent, long invoice, int farmerId, int harvestId)
         {
             ValidateDomain(timestamp, greenWeight, dryWeight, register, price, wholePercent, invoice);
             FarmerId = farmerId;
             HarvestId = harvestId;
         }
 
-        public void ValidateDomain(DateTime timestamp, float? greenWeight, float? dryWeight, int register, float? price, float? wholePercent, long invoice)
+        public void ValidateDomain(DateTime timestamp, decimal? greenWeight, decimal? dryWeight, int register, decimal? price, decimal? wholePercent, long invoice)
         {
             DomainExceptionValidation.When(timestamp == DateTime.MinValue, LoadMessage.InvalidTimestamp);
